requirements.txt
Flask
gunicorn
discord.py>=1.7.3